refactor: modularize health monitor
