+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -195,6 +195,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 680 to 200 lines by extracting core, collector, analyzer, and config modules. Main orchestrator now imports these modules while preserving functionality.
 
+### MODERATE-022: Health Monitor Modularization ✅
+- **File**: `src/core/health/health_monitor.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-24
+- **Summary**: Split monolithic health monitor into core, alerts, and reports modules and introduced an orchestrator coordinating them.
+
 
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
